+++ src/Commands/CreatePermission.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Spatie\Permission\Commands;
+namespace Daniilborovkov\Permission\Commands;
 
 use Illuminate\Console\Command;
-use Spatie\Permission\Contracts\Permission as PermissionContract;
+use Daniilborovkov\Permission\Contracts\Permission as PermissionContract;
 
 class CreatePermission extends Command
 {