+++ src/Middlewares/PermissionMiddleware.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Spatie\Permission\Middlewares;
+namespace Daniilborovkov\Permission\Middlewares;
 
 use Closure;
-use Spatie\Permission\Exceptions\UnauthorizedException;
+use Daniilborovkov\Permission\Exceptions\UnauthorizedException;
 
 class PermissionMiddleware
 {